Big cleanup and adding of convenient macro settings with assert fix

+++ src/memory/libc.h
@@ -17,7 +17,7 @@
  *
  * @return a pointer to the allocated memory
  */
-void *stc_mem_libc_alloc(void *ctx, u64 size);
+void *stc_mem_libc_alloc(void *ctx, size_t size);
 
 /**
  * Frees memory of given size using LibC (free).
@@ -26,7 +26,7 @@ void *stc_mem_libc_alloc(void *ctx, u64 size);
  * @param[in] p the pointer to the memory to free
  * @param[in] size the size of memory to free (ignored)
  */
-void stc_mem_libc_free(void *ctx, void *p, u64 size);
+void stc_mem_libc_free(void *ctx, void *p, size_t size);
 
 /**
  * Creates a pointer to a statically defined memory manager that uses LibC.
